step 2
we define global variables for paper, rock and scissors.
create function for get user choice with prompt, wrote a condition with if for importing valid amount by user.
of course we return each part of function separately...
in the end we connect our function with addEventListener to button,
event is Click and we use an anonymous function.
first we greet user in console then execute the playerChoice function